#384 Support adding optional translations of company name
