+++ public/resources/core-company-config.json
@@ -4,7 +4,8 @@
         "name": {
             "en": "Basic",
             "ko": "기본",
-            "zh": "基本"
+            "zh-Hans": "基本",
+            "zh-Hant": "基本"
         }
     },
     {
